+++ tests/pytests/conftest.py
@@ -1,0 +1,6 @@
+import os
+import sys
+
+# add src to sys.path
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
+print(sys.path)
